fix: get methods
Added get methods for artist and title. YouTube-DL won't necessarly be 
able to get all data for any given video, so there needs to be a get 
method to get the attributes in the preferred priority. ie., first try 
the artist attribute, but if it's not set, then fallback to the video 
creator.