feat: overhaul of song download flow
The previous download form required the user to find all the information
for themselves. This is inconvenient and error prone. To make the
process smoother, the new process allows the user to search for metadata
with the Spotify search API and to search for audio through the YouTube
search API.

The download process on the server side has also been changed to provide
more detailed information about the download progress. Previously the
user was only able to see the progress of the download step, but the
download step is significantly quicker than the extraction step. YT-DLP
does not provide progress information on the extraction step. The
process has now been split into two separate steps: download and
extract. By calling ffmpeg directly, the progress can be monitored and
relayed to the end user.